refactor(governance): split vote card render sections
