+++ apps/governance.mento.org/app/components/voting/vote-card.tsx
@@ -3,7 +3,8 @@ import { TransactionLink } from "@/components/proposal/components/TransactionLin
 import { Timer } from "@/components/timer";
 import { deriveVoteCardState } from "@/components/voting/derive-vote-card-state";
 import { getGovernanceTransactionErrorMessage } from "@/components/voting/get-governance-transaction-error-message";
-import { VoteCardCancelActions } from "@/components/voting/vote-card-cancel-actions";
+import { VoteCardActions } from "@/components/voting/vote-card-actions";
+import { VoteCardSpecialContent } from "@/components/voting/vote-card-special-content";
 import { useDelayedVoteCardRefire } from "@/components/voting/use-delayed-vote-card-refire";
 import { getWatchdogMultisigAddress } from "@/config";
 import { useLocksByAccount } from "@/contracts";
@@ -22,19 +23,16 @@ import { getTimelockOperationId } from "@/contracts/governor/utils/get-timelock-
 import { Proposal, ProposalState } from "@/graphql/subgraph/generated/subgraph";
 import { useVeMentoDelegationSummary } from "@/hooks/use-ve-mento-delegation-summary";
 import {
-  Button,
   Card,
   CardContent,
   CardDescription,
   CardHeader,
   CardTitle,
-  IconLoading,
 } from "@repo/ui";
-import { ConnectButton, NumbersService } from "@repo/web3";
+import { NumbersService } from "@repo/web3";
 import { useAccount, useChainId } from "@repo/web3/wagmi";
 import * as Sentry from "@sentry/nextjs";
 import { CheckCircle2, CircleCheck, XCircle, XCircleIcon } from "lucide-react";
-import Link from "next/link";
 import { useEffect, useMemo, useRef, useState } from "react";
 import spacetime from "spacetime";
 import { formatUnits, keccak256, toHex } from "viem";
@@ -690,366 +688,6 @@ export const VoteCard = ({
   const isVotedForAbstain = voteReceipt?.support === 2;
   const isVotedForReject = voteReceipt?.support === 0;
   const hasVoted = voteReceipt?.hasVoted;
-  const usedVoteOptionButtonLocator = "usedVoteOptionButton";
-
-  // Render vote buttons based on state
-  const renderActions = () => {
-    switch (currentState) {
-      case "loading":
-      case "confirming":
-      case "signing":
-        return null;
-
-      case "insufficient-mento":
-        return (
-          <Button variant="default" size="lg" clipped="default" asChild>
-            <Link href="/voting-power">Lock MENTO Tokens</Link>
-          </Button>
-        );
-
-      case "voted":
-      case "defeated":
-      case "expired":
-      case "finished":
-        // Show disabled buttons for finished proposals
-        if (hasVoted) {
-          return (
-            <div className="gap-4 md:grid-cols-3 grid grid-cols-1">
-              {isVotedForApprove && (
-                <Button
-                  variant="approve"
-                  size="lg"
-                  data-testid={usedVoteOptionButtonLocator}
-                  disabled
-                >
-                  Your vote: YES
-                </Button>
-              )}
-              {isVotedForAbstain && (
-                <Button
-                  variant="abstain"
-                  size="lg"
-                  data-testid={usedVoteOptionButtonLocator}
-                  disabled
-                >
-                  Your vote: Abstain
-                </Button>
-              )}
-              {isVotedForReject && (
-                <Button
-                  variant="reject"
-                  size="lg"
-                  data-testid={usedVoteOptionButtonLocator}
-                  disabled
-                >
-                  Your vote: NO
-                </Button>
-              )}
-            </div>
-          );
-        }
-
-        return null;
-
-      case "queued": {
-        // canExecute derives from isVetoPeriodOver (ticks every second) so the
-        // CTA flips in sync with the header countdown when ETA passes.
-        const canExecute = queueEndTime && isVetoPeriodOver;
-
-        if (!address) {
-          return (
-            <div className="col-span-full flex justify-center">
-              <ConnectButton
-                size="lg"
-                text={
-                  canExecute ? "Connect Wallet to Execute" : "Proposal Queued"
-                }
-                fullWidth
-                disabled={!canExecute}
-              />
-            </div>
-          );
-        }
-        if (canExecute) {
-          return (
-            <div className="gap-4 flex flex-col">
-              <Button
-                variant="default"
-                size="lg"
-                clipped="default"
-                onClick={handleExecute}
-                disabled={isAwaitingExecuteSignature || isExecuteConfirming}
-                data-testid="executeProposalButton"
-                className="w-full"
-              >
-                {isAwaitingExecuteSignature
-                  ? "Confirm in Wallet"
-                  : isExecuteConfirming
-                    ? "Executing..."
-                    : "Execute Proposal"}
-              </Button>
-              <VoteCardCancelActions
-                watchdogCancelAction={watchdogCancelAction}
-              />
-            </div>
-          );
-        }
-        // Show disabled button during veto period or cancel button for watchdog
-        return (
-          <div className="gap-4 flex flex-col">
-            <Button
-              variant="default"
-              size="lg"
-              clipped="default"
-              disabled
-              className="w-full"
-            >
-              In Veto Period
-            </Button>
-            <VoteCardCancelActions
-              watchdogCancelAction={watchdogCancelAction}
-            />
-          </div>
-        );
-      }
-
-      case "executed": {
-        // Show link to execution transaction if available
-        const executionTxHash = proposal.proposalExecuted?.[0]?.transaction?.id;
-        return (
-          <div className="flex justify-center">
-            {executionTxHash ? (
-              <Button variant="outline" size="lg" asChild>
-                <TransactionLink txHash={executionTxHash} className="w-full">
-                  View Execution Transaction
-                </TransactionLink>
-              </Button>
-            ) : (
-              <Button variant="default" size="lg" disabled>
-                Proposal Executed
-              </Button>
-            )}
-          </div>
-        );
-      }
-
-      case "succeeded":
-        // Show queue button for succeeded proposals
-        if (!address) {
-          return (
-            <div className="col-span-full flex justify-center">
-              <ConnectButton
-                size="lg"
-                text="Connect Wallet to Queue"
-                fullWidth
-              />
-            </div>
-          );
-        }
-        return (
-          <div className="gap-4 flex flex-col">
-            <Button
-              variant="default"
-              size="lg"
-              clipped="default"
-              onClick={handleQueue}
-              disabled={isAwaitingQueueSignature || isQueueConfirming}
-              data-testid="queueProposalButton"
-              className="w-full"
-            >
-              {isAwaitingQueueSignature
-                ? "Confirm in Wallet"
-                : isQueueConfirming
-                  ? "Queueing..."
-                  : "Queue for Execution"}
-            </Button>
-            <VoteCardCancelActions
-              watchdogCancelAction={watchdogCancelAction}
-              proposerCancelAction={proposerCancelAction}
-            />
-          </div>
-        );
-
-      case "pending":
-        // Show disabled buttons for pending proposals
-        return (
-          <>
-            <div className="gap-4 md:grid-cols-3 grid grid-cols-1">
-              <Button variant="approve" size="lg" disabled>
-                Voting Not Started
-              </Button>
-              <Button variant="abstain" size="lg" disabled>
-                Voting Not Started
-              </Button>
-              <Button variant="reject" size="lg" disabled>
-                Voting Not Started
-              </Button>
-            </div>
-            <VoteCardCancelActions
-              proposerCancelAction={proposerCancelAction}
-            />
-          </>
-        );
-
-      case "ready":
-        if (!address) {
-          return (
-            <div className="col-span-full flex justify-center">
-              <ConnectButton
-                size="lg"
-                text="Connect Wallet to Vote"
-                fullWidth
-              />
-            </div>
-          );
-        }
-        return (
-          <>
-            <div className="gap-4 md:grid-cols-3 grid grid-cols-1">
-              <Button
-                variant="approve"
-                size="lg"
-                onClick={() => handleVote(1)}
-                disabled={
-                  proposalState !== ProposalState.Active ||
-                  isDeadlinePassed ||
-                  isAwaitingUserSignature ||
-                  isConfirming
-                }
-                data-testid="yesProposalButton"
-              >
-                Vote YES
-              </Button>
-              <Button
-                variant="abstain"
-                size="lg"
-                onClick={() => handleVote(2)}
-                disabled={
-                  proposalState !== ProposalState.Active ||
-                  isDeadlinePassed ||
-                  isAwaitingUserSignature ||
-                  isConfirming
-                }
-                data-testid="abstainProposalButton"
-              >
-                Abstain
-              </Button>
-              <Button
-                variant="reject"
-                size="lg"
-                onClick={() => handleVote(0)}
-                disabled={
-                  proposalState !== ProposalState.Active ||
-                  isDeadlinePassed ||
-                  isAwaitingUserSignature ||
-                  isConfirming
-                }
-                data-testid="noProposalButton"
-              >
-                Vote NO
-              </Button>
-            </div>
-            <VoteCardCancelActions
-              proposerCancelAction={proposerCancelAction}
-            />
-          </>
-        );
-
-      default:
-        return null;
-    }
-  };
-
-  // Get loading text based on state
-  const getLoadingText = () => {
-    switch (currentState) {
-      case "loading":
-        return "Loading voting information...";
-      case "signing":
-        if (isAwaitingExecuteSignature) {
-          return "Waiting for execution confirmation...";
-        }
-        if (isAwaitingQueueSignature) {
-          return "Waiting for queue confirmation...";
-        }
-        return "Waiting for confirmation...";
-      case "confirming":
-        if (isExecuteConfirming) {
-          return "Proposal is being executed";
-        }
-        if (isQueueConfirming) {
-          return "Proposal is being queued";
-        }
-        return "Your vote is being processed";
-      default:
-        return "";
-    }
-  };
-
-  // Special content for certain states
-  const renderSpecialContent = () => {
-    const loadingStates = ["loading", "signing", "confirming"];
-
-    if (loadingStates.includes(currentState)) {
-      return (
-        <div className="gap-4 flex flex-col items-center">
-          <IconLoading />
-          <p
-            className="text-muted-foreground"
-            data-testid={
-              currentState === "signing" && "waitingForConfirmationLabel"
-            }
-          >
-            {getLoadingText()}
-          </p>
-          {currentState === "signing" &&
-            !isAwaitingExecuteSignature &&
-            !isAwaitingQueueSignature && (
-              <p
-                className="text-sm text-muted-foreground"
-                data-testid="waitingForConfirmationDescriptionLabel"
-              >
-                You are voting{" "}
-                {variables?.args?.[1] === 1
-                  ? "YES"
-                  : variables?.args?.[1] === 0
-                    ? "NO"
-                    : "ABSTAIN"}{" "}
-                on this proposal
-              </p>
-            )}
-          {currentState === "signing" && isAwaitingExecuteSignature && (
-            <p
-              className="text-sm text-muted-foreground"
-              data-testid="waitingForExecutionDescriptionLabel"
-            >
-              You are executing this proposal
-            </p>
-          )}
-          {currentState === "signing" && isAwaitingQueueSignature && (
-            <p
-              className="text-sm text-muted-foreground"
-              data-testid="waitingForQueueDescriptionLabel"
-            >
-              You are queueing this proposal for execution
-            </p>
-          )}
-          {currentState === "confirming" &&
-            currentTxHash &&
-            (hash || executeHash || queueHash) && (
-              <Button variant="outline" size="sm" asChild className="mt-2">
-                <TransactionLink txHash={currentTxHash}>
-                  View on explorer
-                </TransactionLink>
-              </Button>
-            )}
-        </div>
-      );
-    }
-
-    return null;
-  };
-
   const quorumLabel = useMemo(() => {
     let label = "";
 
@@ -1151,7 +789,18 @@ export const VoteCard = ({
             : "space-y-8 p-4 xl:p-8"
         }
       >
-        {renderSpecialContent()}
+        <VoteCardSpecialContent
+          currentState={currentState}
+          isAwaitingExecuteSignature={isAwaitingExecuteSignature}
+          isAwaitingQueueSignature={isAwaitingQueueSignature}
+          isExecuteConfirming={isExecuteConfirming}
+          isQueueConfirming={isQueueConfirming}
+          voteSupport={variables?.args?.[1] as number | undefined}
+          currentTxHash={currentTxHash}
+          hash={hash}
+          executeHash={executeHash}
+          queueHash={queueHash}
+        />
 
         {!["loading", "confirming", "signing"].includes(currentState) && (
           <>
@@ -1184,7 +833,30 @@ export const VoteCard = ({
                   : "mt-8"
               }
             >
-              {renderActions()}
+              <VoteCardActions
+                currentState={currentState}
+                address={address}
+                proposal={proposal}
+                proposalState={proposalState}
+                hasVoted={hasVoted}
+                isVotedForApprove={isVotedForApprove}
+                isVotedForAbstain={isVotedForAbstain}
+                isVotedForReject={isVotedForReject}
+                queueEndTime={queueEndTime}
+                isVetoPeriodOver={isVetoPeriodOver}
+                isDeadlinePassed={isDeadlinePassed}
+                isAwaitingUserSignature={isAwaitingUserSignature}
+                isConfirming={isConfirming}
+                isAwaitingExecuteSignature={isAwaitingExecuteSignature}
+                isExecuteConfirming={isExecuteConfirming}
+                isAwaitingQueueSignature={isAwaitingQueueSignature}
+                isQueueConfirming={isQueueConfirming}
+                onExecute={handleExecute}
+                onQueue={handleQueue}
+                onVote={handleVote}
+                watchdogCancelAction={watchdogCancelAction}
+                proposerCancelAction={proposerCancelAction}
+              />
             </div>
 
             {activeTransactionErrorMessage &&
